requirements.txt
streamlit
langchain
python-dotenv
langchain_community
chromadb
huggingface_hub